% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrationDistributions.R
\name{assumption1}
\alias{assumption1}
\alias{assumption2}
\alias{assumption3}
\title{Functions for Assumptions}
\usage{
assumption1(dist)

assumption2(dist)

assumption3(dist)
}
\arguments{
\item{dist}{The distribution parameter.}

\item{model}{The model object.}

\item{...}{Additional arguments to be passed to other functions.}
}
\value{
A function with specified assumption for the model.
}
\description{
The functions assumption1, assumption2, and assumption3 implement different assumptions for a model, namely A.I, A.II', and A.II'', respectively.
}
\details{
These functions implement different assumptions for a model.


Each function assigns the specified assumption to the parent environment and processes predefined measures if available.
}
