% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelStructures.R
\name{glm_to_compint}
\alias{glm_to_compint}
\alias{logistf_to_compint}
\title{Changing-model-structures: Functions for transforming specific model objects to CompInt model objects}
\usage{
glm_to_compint(model)

logistf_to_compint(model, data)
}
\description{
... Complete documentation !!! #TOFIX
}
\section{Functions}{
\itemize{
\item \code{glm_to_compint()}: Transforms a standard glm model into a CompInt model

\item \code{logistf_to_compint()}: Transforms a logistf model into a CompInt model

}}
