% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompInt-package.R
\docType{package}
\name{CompInt-package}
\alias{CompInt}
\alias{CompInt-package}
\title{CompInt: Compute comparable and interpretable reporting quantities for different model objects}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package is in the development phase. Eventually, it will contain implementations of all quantities from the preprint of Kümpel & Hoffmann (2022).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://11annah.github.io/CompInt/}
  \item \url{https://github.com/11annah/CompInt}
  \item Report bugs at \url{https://github.com/11annah/CompInt/issues}
}

}
\author{
\strong{Maintainer}: Hannah Schulz-Kümpel \email{hannah.kuempel@stat.uni-muenchen.de} (\href{https://orcid.org/0000-0003-3972-8392}{ORCID})

Other contributors:
\itemize{
  \item Pat Callahan [contributor]
  \item Julia Höpler [contributor]
}

}
\keyword{internal}
