% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelStructures.R
\name{draws_from_paramdist}
\alias{draws_from_paramdist}
\title{Draws samples from the parameter distribution of a CompInt_model object.}
\usage{
draws_from_paramdist(model, ndraws = 1000, seed = NULL, ...)
}
\arguments{
\item{ndraws}{Number of draws to generate.}

\item{seed}{Seed for the random number generator.}

\item{...}{Additional arguments passed to the underlying draw function. This function is stored in model$pseudo_posterior$fun.}

\item{object}{A CompInt_model object.}
}
\value{
A matrix of draws from the parameter distribution.
}
\description{
This function generates draws from the parameter distribution of a CompInt_model object.
}
\details{
#TOFIX!!!
}
\examples{
\dontrun{
# Example usage
glm <- ###some GLM needs to be fit!!! #TOFIX!!!
model <- glm_to_compint(glm)
draws <- draws_from_paramdist(model, ndraws = 1000, seed = 123)
}

#TOFIX: is importFrom needed here? I mean it depends on the pseudo_posterior fun

}
