% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrameworkQuantity_gMEs.R
\name{get_gME}
\alias{get_gME}
\title{Calculate Generalized Marginal Effects (gME)}
\usage{
get_gME(
  model_fit,
  reg_of_interest = NULL,
  integration = NULL,
  seed = NULL,
  ndraws = 1000,
  separate_interactions = FALSE,
  catRIbin = FALSE,
  ...
)
}
\arguments{
\item{model_fit}{A fitted model object.}

\item{reg_of_interest}{The regressor of interest for which gME values are calculated.}

\item{integration}{An integration function. Default is NULL. #TOFIX ... \code{\link{assumption1}} \code{\link{assumption2}} \code{\link{assumption3}}}

\item{seed}{Seed for random number generation. Default is NULL.}

\item{ndraws}{Number of draws for sampling. Default is 1000.}

\item{separate_interactions}{Boolean indicating whether interactions should be separated. Default is FALSE.}

\item{catRIbin}{Boolean indicating categorical RI binning. Default is FALSE.}

\item{...}{Additional arguments to be passed to other functions.}
}
\value{
A matrix containing gME values for the regressor of interest.
}
\description{
This function computes generalized Marginal Effects (gME) for a given model and regressor of interest.
}
\details{
This function computes gME values based on the specified model and regressor of interest. It supports various types of models including GLM and GLMM. The gME values are computed either empirically or through other standard options depending on the distribution.
}
\note{
...
}
\examples{
\dontrun{
model <- lm(mpg ~ cyl + disp + hp + drat, data = mtcars)
get_gME(model_fit = model, reg_of_interest = "cyl")
}

}
\references{
Kümpel, H. & Hoffmann, S. A formal framework for generalized reporting methods in parametric settings. ArXiv Prepr. ArXiv221102621 (2022).
}
