% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_theta.R
\name{make_linear_predictor}
\alias{make_linear_predictor}
\title{Extracts the linear predictor from a model formula.}
\usage{
make_linear_predictor(
  mod,
  reg_of_interest = NULL,
  separate_interactions = FALSE
)
}
\arguments{
\item{mod}{A model object of class "CompInt".}

\item{reg_of_interest}{The regressor of interest. If specified, the original regressor name in the linear predictor will be replaced with "RI".}

\item{separate_interactions}{Logical, indicating whether interaction terms should be denoted and treated as a separate regressor.}
}
\value{
A character string representing the linear predictor.
}
\description{
This function extracts the linear predictor from a model formula and formats it as a string.
}
\details{
The function retrieves the coefficients from the model object and constructs the linear predictor based on the model formula.
#TOFIX!!!
}
\examples{
\dontrun{
# Example usage
glm <- ###some GLM needs to be fit!!! #TOFIX!!!
model <- glm_to_compint(glm)
linear_pred <- make_linear_predictor(model, reg_of_interest = "TOFIX", separate_interactions = TRUE)
}


}
