% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_theta.R
\name{produce_model_definition}
\alias{produce_model_definition}
\title{#TOFIX}
\usage{
produce_model_definition(
  mod,
  separate_interactions = FALSE,
  name_targetvar = NULL
)
}
\arguments{
\item{mod}{A model object of class "CompInt".}

\item{separate_interactions}{Logical, indicating whether interaction terms should be denoted and treated as a separate regressor.}
}
\value{
A character string representing the linear predictor.
}
\description{
#TOFIX
}
\details{
#TOFIXThe function retrieves the coefficients from the model object and constructs the linear predictor based on the model formula.
#TOFIX!!!
}
\examples{
\dontrun{
# Example usage
#TOFIXglm <- ###some GLM needs to be fit!!! #TOFIX!!!
#TOFIXmodel <- glm_to_compint(glm)
#TOFIXlinear_pred <- produce_model_definition(model, reg_of_interest = "TOFIX", separate_interactions = TRUE)
}


}
